predictor/requirements.txt
# =========================
# Core framework
# =========================
Django>=4.2,<5.0

# =========================
# Deep Learning
# =========================
tensorflow>=2.13,<2.16
keras>=2.13,<2.16

# =========================
# Image processing
# =========================
opencv-python>=4.8
Pillow>=9.5

# =========================
# Scientific stack
# =========================
numpy>=1.23,<2.0
scipy>=1.9
matplotlib>=3.7

# =========================
# Model explainability (Grad-CAM)
# =========================
scikit-learn>=1.2

# =========================
# Utility
# =========================
tqdm>=4.65
pandas>=2.0

# Vision / NLP
transformers>=4.36

tf-keras>=2.15
